+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## TBA - 0.1.1
 
 - Add GitHub actions example and instructions to validate geofeed using CLI hook command
+- Add `geofeed-tools dump --format {json,csv,table}` for JSON, geofeed CSV, and tabulated output modes
 
 ## 2026-05-26 - 0.1.0